feat: DX12 renderer, pit lane FSM, race results, FBX loader, telemetry validation, AI opponent improvements
- X-Engine: DX12 renderer with shaders, win32 windowing, new tests
- Engine: pit_stop_fsm, pit_service_unit, race_manager, result_storage, standings improvements
- Audio: audio engine with spatial sound
- Debug: debug_agent and debug_console
- Tracking: checkpoint, lap_system, simulated_gps, tracked_telemetry improvements
- UI: hud and menu_system updates
- Vehicle: car_model parameter refinements
- Gameplay: refactoring and input debouncing
- Tools: track_diagram improvements, FBX conversion scripts
- Tests: new checkpoint_validation, race_data_collection, telemetry_validation test suites
- Unity: URP package installer, InputCompat, TrackGenerator updates
- FBX mesh loader, JSON parser

All 558 tests passing.

+++ engine/simulation/simulation.h
@@ -15,7 +15,7 @@ namespace p0::simulation {
 // Simulation configuration
 struct SimulationParams {
   double dt = 1.0 / 60.0;                // s, base timestep (60 Hz)
-  double substeps = 4.0;                 // number of physics sub-steps per frame
+  int substeps = 4;                      // number of physics sub-steps per frame
   bool use_abs = true;                   // enable ABS (placeholder)
   bool use_tcs = true;                   // enable TCS (placeholder)
 };
